Add clock dynamics types
Adds Clock and VariableClock dynamics types, both of which use Durations to exactly represent time.
Using Duration avoids floating-point issues that crop up when using Linear or Polynomial resource to represent time,
as well as not needing conversions to and from Duration.

Of particular note are the stopwatch-style effects and resource-level comparison functions for VariableClock.
Combining these with `DiscreteResources.when` and `Reactions` is especially useful for expressing time-based conditions and behaviors.