Add approximation utilities
Adds utilities for constructing approximations.

These are particularly useful for Unstructured resources with Double as a value,
but some approximations generalize beyond this.

Approximation is broken into several stages:
 1. Choose an approximation type.
    The ready-made choices here are "discrete", "secant", or "Taylor" approximations, but the interface permits others to be defined later.
 2. Choose a strategy for running those approximations - choosing a divergence estimator or interval function, depending on the approximation type.
    There are ready-made strategies for using uniformly-spaced samples or attempting to bound the final error.
 3. If error-bounding is used, choose an error tolerance and error measurement.
    Again, there are several ready-made options, based on both direct methods and analytic estimates, depending on the available information.
    Further parameters may be needed to configure these error estimates.

Breaking the problem down like this allows for more focused testing, as well as component re-use.
Doing approximation on an Unstructured resource allows the value of the result to influence the sampling strategy -
rather than heuristically choosing a sampling strategy to get "good enough" data, we can measure the result to choose samples.
This can be an efficiency boon when the approximated resource is cheap, but downstream resources or tasks are expensive.