Take expiry into account for approximation interval
When solving for an interval over which to take an approximation, we take the expiry of the dynamics
to be approximated into account. In the case when no solution is found, we should still take expiry into account.

This is especially important when that expiry is much shorter than the typical approximation interval.
In this case, we can often choose the full expiry as the approximation interval without reaching our maximum error tolerance,
resulting in a NoBracketingException from the solver. In this case, we should choose the expiry, not the maximum approximation interval,
to be consistent with the time range we searched.