WIP - Add three examples of activity modeling with expansion
Adds three examples of how activities might be modeled in a way that includes / is centered around command expansion.

In the "shallow" approach, commands are completely unmodeled.
The activity builds up SEQ JSON, with the help of an in-memory object model, and reports the expansion.
All modeling is done by the activity.

In the "moderate" approach, commands exist as Java classes, and are used to write the sequence.
However, those command classses are skeletons with little to no behavior, and activities may or may not "run" the sequence they build.
Most modeling is done by the activity.

In the "deep" appraoch, commands are full-fledged Aerie activities, and include their effects on the model.
Activities build sequences with these commands, and run those sequences to model most of their behavior.
Most modeling is done by the commands themselves, though activity-level modeling is still possible.
Additionally, we demonstrate facilities for activity-level modeling to "watch" or "synchronize with" the sequence execution.

Finally, this project demonstrates that these approaches can be mix-and-matched, including deeper approaches being called by shallower approaches.
All three levels of "science" activities call the Warm_Up activity, which is deeply modeled.