WIP - Generalize DiscreteResourceBuilder
Generalizes the DiscreteResourceBuilder to a MutableResourceBuilder, and strips registration out of this builder.
Since derived resources will need to be registered separately anyways, it seems more consistent to register mutable resources the same way.
The builders now use Naming.name to attach the name they're given to the resource they construct.
If the registrar consults Naming.name (or at least can consult it), then we don't need special builder constructs to avoid repeating the name there.

At the moment, I'm leaning towards just accepting that the registrar will ask users to specify a value mapper for all discrete resources, even where this duplicates what they've given to the builder.
It may be worth having an "auto value mapper" overload for the registrar, which pulls the initial value of the resource, and if it's a primitive, uses the appropriate value mapper automatically.
This would avoid a lot of the value mapper specifications, and would also work on derived resources, at the cost of being a little janky under the hood.

Another approach I'm considering is "registering" a value mapper, akin to what we do with names. This doesn't address the derived resources though, and seems unnecessarily brittle.