WIP - Fixing compile errors
Fixing compile errors, mostly around supplying names and value mappers where needed.
This opened a can of worms around using the merlin processor in contrib - it currently depends on contrib, so trying to use it creates a circular dependency.
OTOH, writing mappers for all these record types feels wrong, given we have AutoValueMapper for records.
We could use mapping and isomorphism to short-cut some of this, but it would lead to an inconsistency in how we save out the results, which I don't love.