Debugging tests around streamline init
Refactors the init process for streamline to go through StreamlineSystem.init, a purpose-built entry point to centralize singleton construction.
This also made it convenient to build the registrar as a singleton, but did require allowing any singleton to be re-initialized to support unit tests.

The SimulationSystem class also provides a good place to put global entities like the simulation clock which don't fit well anywhere else.