Add CompactSupport resources
Adds CompactSupport resources, a kind of MutableResource lens which allows for infinite or infeasibly large sets of logical resources,
provided there's a compact, feasibly large set with a non-default value.

This was written to support an experiment translating PDDL to Aerie, where we'd like to represent functions which return resources,
and we either don't know or can't feasibly enumerate the full domain of the function to statically declare all of those resources.